# swagger.yaml
...
swagger: "2.0"
info:
  description: |
    
  version: "1.0.0"
  title: "Authentication"
  termsOfService: ""
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /auth/forgot-password:
    post:
      tags:
      - Authentication
      summary: "Forgot Password"
      description: "Forgot Password"
      operationId: "Forgot Password"
      consumes:
      - "application/json"
      produces:
      - "applicaion/json"
      parameters:
      - in: "body"
        name: "body"
        description: Forgot Password
        required: true
        schema:
          $ref: "#/definitions/ForgotPassword"
      responses:
        200:
          description: "Response on successful request"
      x-swagger-router-controller: "User"
  /auth/change-forgot-password:
    post:
      tags:
      - Authentication
      summary: "Forgot Password"
      description: "Forgot Password"
      operationId: "Forgot Password"
      consumes:
      - "application/json"
      produces:
      - "applicaion/json"
      parameters:
      - in: "body"
        name: "body"
        description: Forgot Password
        required: true
        schema:
          $ref: "#/definitions/ChangeForgotPassword"
      responses:
        200:
          description: "Response on successful request"
      x-swagger-router-controller: "User"
...
definitions:
...
  ForgotPassword:
    type: object
    required:
    - "email"
    properties:
      email:
        type: "string"
        description: Fogot Password Email
  ChangeForgotPassword:
    type: object
    required:
    - "code"
    - "password"
    properties:
      code:
        type: "string"
        description: Fogot Password Code
      password:
        type: "string"
        description: New Password
...